// src/enums.ts
export enum TripEnvisions {
  ROMANCE = "Romance",
  PARTY = "Party",
  DOING_BUSINESS = "Doing Busineess",
  WORK_REMOTELY = "Work Remotely",
  NATURE_EXPLORATION = "Nature Exploration",
  CITY_CRUISING = "City Cruising",
  BEACH = "Beach",
  CULTURAL_EXPERIENCE = "Cultural Experience",
  SPORT_ACTIVITY = "Sport Activity",
  WELLNESS_AND_RELAXATION = "Welleness and Relax"
}

export enum BudgetAmounts {
  DONT_KNOW = "Dont know",
  FROM_0_TO_100_EUROS = "From 0 to 100 euros",
  FROM_100_TO_200_EUROS = "From 100 to 200 euros",
  FROM_200_TO_500_EUROS = "From 200 to 500 euros",
  FROM_500_EUROS = "From 500 euros"
}